+++ packages/react-storage/src/components/StorageBrowser/views/LocationActionView/DeleteView/utils.ts
@@ -1,0 +1,6 @@
+import { DeleteHandlerData } from '../../../actions';
+import { Task } from '../../../tasks';
+
+export const getFolderText = ({
+  data: { fileKey, key },
+}: Task<DeleteHandlerData>): string => key.slice(0, -fileKey.length);